+++ src/components/Admin/pages/AdminAgendas/index.tsx
@@ -1,0 +1,8 @@
+
+export default function AdminAgendas() {
+  return (
+    <>
+      <h1>Admin Agendas</h1>
+    </>
+  );
+}
