+++ src/components/Admin/pages/AdminHome/index.tsx
@@ -1,0 +1,8 @@
+
+export default function AdminHome() {
+  return (
+    <>
+      <h1>Admin Home</h1>
+    </>
+  );
+}
