+++ src/components/Coberturas/index.tsx
