+++ src/components/LandingPage/styles.css
@@ -1,0 +1,353 @@
+.img_header {
+  position: relative;
+
+  & img {
+    width: 100%;
+    height: 300px;
+  }
+}
+
+.div-mobile {
+  position: absolute;
+  z-index: 100;
+  height: 100%;
+  top: 0%;
+  width: 00%;
+  overflow: hidden;
+  background-color: #2e2b72;
+  transition: 0.3s;
+}
+.abri-menu {
+  width: 70%;
+  overflow: visible;
+  transition: 0.3s;
+}
+.Xmark {
+  font-size: 25px;
+  color: #e31c80;
+  right: 20px;
+  top: 15px;
+  position: absolute;
+}
+
+.nav_mobile {
+  height: 100%;
+  padding-top: 100px;
+}
+
+.nav_mobile ul {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  list-style: none;
+  height: 100%;
+  padding-left: 0px;
+  position: relative;
+
+  & li {
+    padding: 5px;
+    padding-left: 24px;
+  }
+  & li:hover {
+    background-color: #e31c80;
+  }
+
+  & a {
+    color: white;
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    text-decoration: none;
+  }
+}
+.nav_mobile ul::before {
+  content: "";
+  height: 100%;
+  width: 14px;
+  position: absolute;
+  left: 0;
+  background-color: #e31c80;
+}
+
+.containerCpl {
+  position: absolute;
+  left: 50%;
+  top: 6%;
+  transform: translate(-50%, -45%);
+}
+
+.contNet {
+  margin-top: -200px;
+  position: relative;
+}
+
+.min_nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.about {
+  display: flex;
+  gap: 1rem;
+  margin-left: 300px;
+  color: #ff8b00;
+}
+
+.img_min_nav {
+  display: flex;
+  align-items: center;
+  gap: 20px;
+}
+
+.imagem {
+  display: flex;
+  align-items: center;
+  gap: 4px;
+}
+
+.img_min_nav .img {
+  & img {
+    width: 25px;
+    height: 25px;
+  }
+}
+
+.imgTwiter {
+  width: 25px;
+  height: 25px;
+  background-color: #44aed5;
+  text-align: center;
+
+  & i {
+    color: white;
+  }
+}
+
+.min_Buriti {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+
+  & div:nth-child(1) {
+    width: 7px;
+
+    & img {
+      width: 100%;
+    }
+  }
+
+  & div:nth-of-type(2) {
+    background-color: #eceef5;
+    padding: 5px;
+    padding-top: 0px;
+    padding-bottom: 0;
+    border: 2px solid #cad4e7;
+  }
+
+  & h3 {
+    font-size: 17px;
+    font-weight: bolder;
+    color: #3d5997;
+    padding-top: 10px;
+
+    & span {
+      font-weight: 400;
+    }
+  }
+
+  & button {
+    padding: 0px 10px 0px;
+    display: flex;
+    align-items: center;
+    gap: 5px;
+    background-color: #eceef5;
+    border: 2px solid #cad4e7;
+    border-radius: 5px;
+    color: #365899;
+
+    & i {
+      padding-bottom: 5px;
+    }
+  }
+}
+
+.logo {
+  display: flex;
+  justify-content: space-between;
+  align-items: start;
+  padding-top: 7px;
+  position: relative;
+
+  & span:nth-child(2) {
+    color: #ff8b00;
+    font-size: 25px;
+  }
+
+  & img {
+    margin-top: -20px;
+  }
+}
+
+.button-usename {
+  height: 35px;
+  width: 400px;
+
+  & input,
+  span {
+    background-color: #e8e8e8 !important;
+    box-shadow: none !important;
+    border: none !important;
+  }
+
+  & button {
+    border: none;
+    background: #505050;
+    color: #b5b5b5;
+    font-size: 14px;
+    padding: 15px;
+    padding-top: 0;
+    padding-bottom: 0;
+  }
+}
+
+.nav_uls ul {
+  position: relative;
+  width: 100%;
+}
+
+.bars {
+  display: none;
+}
+@media screen and (max-width: 1400px) {
+  .containerCpl {
+    padding-right: 20px;
+    padding-right: 11px;
+  }
+  .nav_uls ul::before {
+    width: 98%;
+  }
+}
+
+@media screen and (max-width: 1200px) {
+  .min_nav {
+    padding-top: 10px;
+  }
+  .containerCpl {
+    padding-left: 0;
+  }
+
+  .about {
+    margin-left: 20%;
+  }
+
+  .img_min_nav {
+    gap: 5px;
+  }
+
+  .img_min_nav .img {
+    & img {
+      width: 24px;
+    }
+  }
+
+  .imgTwiter {
+    width: 24px;
+  }
+
+  .min_Buriti {
+    gap: 5px;
+
+    & div:nth-child(1) {
+      width: 7px;
+
+      & img {
+        width: 90%;
+      }
+    }
+
+    & div:nth-of-type(2) {
+      padding: 0px;
+      padding-right: 3px;
+      padding-left: 3px;
+    }
+
+    & button {
+      padding: 0px 8px 0px;
+    }
+  }
+
+  .logo {
+    & span:nth-child(2) {
+      font-size: 18px;
+    }
+  }
+
+  .nav_uls li a {
+    font-size: 12px !important;
+  }
+
+  .contNet {
+    margin-top: -200px;
+  }
+}
+
+@media screen and (max-width: 997px) {
+  .logo {
+    & span:nth-child(2) {
+      display: none;
+    }
+  }
+
+  .nav-bar {
+    display: none;
+  }
+  .rowSlider {
+    display: none;
+  }
+
+  .contNet {
+    margin-top: -160px;
+  }
+
+  .button-usename {
+    display: none;
+  }
+
+  .bars {
+    display: block;
+    font-size: 25px;
+    z-index: 89;
+  }
+
+  .logo_Buriti {
+    padding-right: 10px;
+  }
+  .min_nav {
+    align-items: center;
+    margin-bottom: 10px;
+  }
+  .about {
+    margin-left: 0%;
+  }
+}
+@media screen and (max-width: 768px) {
+  .img_min_nav {
+    display: none;
+  }
+  .min_nav {
+    justify-content: center;
+  }
+}
+
+@media screen and (max-width: 400px) {
+  .min_nav {
+    padding-top: 20px;
+    margin-bottom: 20px;
+  }
+}
+@media screen and (min-width: 1950px) {
+  .coll-4 {
+    width: 30%;
+  }
+}
