+++ src/components/NotFound/index.tsx
@@ -1,0 +1,9 @@
+const NotFound = () => {
+  return (
+    <div>
+      <h1>404 - Not Found</h1>
+    </div>
+  );
+};
+
+export default NotFound;
