+++ src/components/Noticias/index.tsx
