+++ src/main.tsx
@@ -1,16 +1,13 @@
+import "primeicons/primeicons.css";
+import "primereact/resources/primereact.min.css";
+import "primereact/resources/themes/saga-blue/theme.css";
 import React from "react";
 import ReactDOM from "react-dom/client";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Admin from "./components/adminn/admin.tsx";
 import App from "./App.tsx";
 import "./index.css";
+
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App/>} />
-        <Route path="/admin" element={<Admin/>} />
-      </Routes>
-    </BrowserRouter>
+    <App />
   </React.StrictMode>
 );
